allow playing favorite songs after the contest is over
The listen template was changed to be more generic, and now you can pass
in songs that you liked from across the competition. For the time being,
this is only available after the competition. I want to encourage
listening to the mixtapes as they were prepared for at least the
duration of the contest.