set listen preload to auto
This defaults to metadata, but it may fix the issue of pausing a song
and coming back later to resume and hear it cut out after the buffer was
exhausted. My guess is that it is timing out at the nginx level, so
perhaps this will let the browser greedily buffer the whole song.