Use new visualization system
The new visualization system is focused mainly on performance, but also enables more involved visualizations. Before, visualizations were strictly shaders, which meant that all of the cpu computations for each one have to be performed no matter which shader is actually showing. This caused a lot of CPU usage that wasn't necessary. It also meant that more interesting things, such as particles, couldn't really be used. Now every visualization has its own logic, and the calculations are performed only for the one that's showing (using the magic of rxjs and disposables).