reorganize assets and fix precompile
Now the top-level files in the assets directories are targets and
subdirectories contain assets to require. We're also not linking
directly to the vendor scripts and instead providing shells for them in
our assets directory. We also need to set them to be precompiled, which
I think changed in the sprockets-rails upgrade when we bumped major from
2 to 3.