fix mime extensions and use filename
Instead of manually specifying the extension, we should just use
whatever is in `opts[:filename]` (though we have to remove the leading
dot that `File.extname` gives back to us). We also want to make sure we
send the right mime type for mp3 and m4a respectively, and we don't need
to register `zip` because it already exists.